+++ core/src/main/java/org/apache/struts/chain/commands/ActionCommandBase.java
@@ -25,8 +25,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * Simple abstract class which avoids frequent casting to {@code ActionContext}
- * in commands explicitly intended for use with that class.
+ * Simple abstract class which logs the {@code ActionCommand}-class when
+ * executed.
  */
 public abstract class ActionCommandBase implements ActionCommand {
 
@@ -36,7 +36,30 @@ public abstract class ActionCommandBase implements ActionCommand {
     private final Logger log =
         LoggerFactory.getLogger(ActionCommandBase.class);
 
-    // See interface for Javadoc
-    public abstract boolean execute(ActionContext actionContext)
+    /**
+     * Logs and executes the command.
+     * 
+     * @param actionContext The {@code Context} for the current request
+     *
+     * @return TRUE if processing should halt
+     *
+     * @throws Exception On any error
+     */
+    @Override
+    public final boolean execute(ActionContext actionContext) throws Exception {
+        log.debug("Executing {}", getClass().getName());
+        return execute_(actionContext);
+    }
+
+    /**
+     * Executes the command.
+     *
+     * @param actionContext The {@code Context} for the current request
+     *
+     * @return TRUE if processing should halt
+     *
+     * @throws Exception On any error
+     */
+    protected abstract boolean execute_(ActionContext actionContext)
         throws Exception;
 }