+++ core/src/main/java/org/apache/struts/chain/commands/ExceptionCatcher.java
@@ -105,7 +105,8 @@ public class ExceptionCatcher extends ActionCommandBase implements Filter<Action
      *
      * @throws Exception On any error
      */
-    public boolean execute(ActionContext actionCtx)
+    @Override
+    protected boolean execute_(ActionContext actionCtx)
         throws Exception {
         actionCtx.setException(null);